Increase receive buffer size in VideoStream.c
By default sunshine uses 512 KB for the receive buffer.
But on older devices it is possible that the decoder and network
cannot process all received packets. And new packets overwrite
the packets already received but not yet processed.
Because of this, some frames cannot be decoded.
Increase the receive buffer to a larger one: 64MB.
Just to be sure it is enough.